Various small changes
Trying to make the function self-contained. I added some code in "private" (matlab always looks in the folder "private" before looking for other similarly named functions in the path) that are still missing from the published version of the FMAToolbox.

Removed dependencies (hopefully entirely) to the particular file format we use so the function can be called with the lfp-s only (without relying on SetCurrentSession etc).