exclude noisyIntervals from freezing
also add an unfinished comment advising users to pre-smooth their own speed data with the appropriate smoothing window (as we cannot assume their sampling rate). It may be easier for them to estimate a proper threshold if it applies to the inputted data directly (rather than to a smoothed version of those data).